Fixed factorial allowing double as incoming variable (now only int)

+++ Interface.java
@@ -10,6 +10,6 @@ public interface Interface extends Remote {
 	public double divide(double a, double b) throws RemoteException;
 	public double invert(double a) throws RemoteException;
 	public double power(double a, double b) throws RemoteException;
-	public double factorial(double a) throws RemoteException;
+	public double factorial(int a) throws RemoteException;
 	public double sqareroot(double a) throws RemoteException;
-} 
+}
